e2e and canary use same suite for uniswap
