+++ yarn-project/end-to-end/src/index.ts
@@ -2,3 +2,4 @@
 
 export { cliTestSuite } from './canary/cli.js';
 export { browserTestSuite } from './canary/browser.js';
+export { uniswapL1L2TestSuite, UniswapSetupContext } from './canary/uniswap_l1_l2.js';
